backend/main_backend/requirements.txt
aiofiles==25.1.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.12.0
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# ML & Data Science
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
pandas==2.2.0

# NLP 
# nltk==3.8.1
# spacy==3.7.2

pytesseract==0.3.10
Pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
python-dotenv==1.0.0
# requests==2.31.0  
# aiofiles==23.2.1

#report generation
reportlab
weasyprint